test(config): move resolver tests and add codex cli path invalid cases
